agent: move storage device related code into dedicated files
Move storage device related code into dedicated files.
