+++ src/agent/src/storage/block_handler.rs
@@ -1,0 +1,145 @@
+// Copyright (c) 2019 Ant Financial
+// Copyright (c) 2023 Alibaba Cloud
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+use std::fs;
+use std::os::unix::fs::PermissionsExt;
+use std::path::Path;
+use std::str::FromStr;
+
+use anyhow::{anyhow, Context, Result};
+use protocols::agent::Storage;
+use tracing::instrument;
+
+use crate::device::{
+    get_scsi_device_name, get_virtio_blk_pci_device_name, get_virtio_mmio_device_name,
+    wait_for_pmem_device,
+};
+use crate::pci;
+use crate::sandbox::StorageDeviceObject;
+use crate::storage::{common_storage_handler, new_device, StorageContext, StorageHandler};
+#[cfg(target_arch = "s390x")]
+use crate::{ccw, device::get_virtio_blk_ccw_device_name};
+
+#[derive(Debug)]
+pub struct VirtioBlkMmioHandler {}
+
+#[async_trait::async_trait]
+impl StorageHandler for VirtioBlkMmioHandler {
+    #[instrument]
+    async fn create_device(
+        &self,
+        storage: Storage,
+        ctx: &mut StorageContext,
+    ) -> Result<StorageDeviceObject> {
+        if !Path::new(&storage.source).exists() {
+            get_virtio_mmio_device_name(ctx.sandbox, &storage.source)
+                .await
+                .context("failed to get mmio device name")?;
+        }
+        let path = common_storage_handler(ctx.logger, &storage)?;
+        new_device(path)
+    }
+}
+
+#[derive(Debug)]
+pub struct VirtioBlkPciHandler {}
+
+#[async_trait::async_trait]
+impl StorageHandler for VirtioBlkPciHandler {
+    #[instrument]
+    async fn create_device(
+        &self,
+        mut storage: Storage,
+        ctx: &mut StorageContext,
+    ) -> Result<StorageDeviceObject> {
+        // If hot-plugged, get the device node path based on the PCI path
+        // otherwise use the virt path provided in Storage Source
+        if storage.source.starts_with("/dev") {
+            let metadata = fs::metadata(&storage.source)
+                .context(format!("get metadata on file {:?}", &storage.source))?;
+            let mode = metadata.permissions().mode();
+            if mode & libc::S_IFBLK == 0 {
+                return Err(anyhow!("Invalid device {}", &storage.source));
+            }
+        } else {
+            let pcipath = pci::Path::from_str(&storage.source)?;
+            let dev_path = get_virtio_blk_pci_device_name(ctx.sandbox, &pcipath).await?;
+            storage.source = dev_path;
+        }
+
+        let path = common_storage_handler(ctx.logger, &storage)?;
+        new_device(path)
+    }
+}
+
+#[derive(Debug)]
+pub struct VirtioBlkCcwHandler {}
+
+#[async_trait::async_trait]
+impl StorageHandler for VirtioBlkCcwHandler {
+    #[cfg(target_arch = "s390x")]
+    #[instrument]
+    async fn create_device(
+        &self,
+        mut storage: Storage,
+        ctx: &mut StorageContext,
+    ) -> Result<StorageDeviceObject> {
+        let ccw_device = ccw::Device::from_str(&storage.source)?;
+        let dev_path = get_virtio_blk_ccw_device_name(ctx.sandbox, &ccw_device).await?;
+        storage.source = dev_path;
+        let path = common_storage_handler(ctx.logger, &storage)?;
+        new_device(path)
+    }
+
+    #[cfg(not(target_arch = "s390x"))]
+    #[instrument]
+    async fn create_device(
+        &self,
+        _storage: Storage,
+        _ctx: &mut StorageContext,
+    ) -> Result<StorageDeviceObject> {
+        Err(anyhow!("CCW is only supported on s390x"))
+    }
+}
+
+#[derive(Debug)]
+pub struct ScsiHandler {}
+
+#[async_trait::async_trait]
+impl StorageHandler for ScsiHandler {
+    #[instrument]
+    async fn create_device(
+        &self,
+        mut storage: Storage,
+        ctx: &mut StorageContext,
+    ) -> Result<StorageDeviceObject> {
+        // Retrieve the device path from SCSI address.
+        let dev_path = get_scsi_device_name(ctx.sandbox, &storage.source).await?;
+        storage.source = dev_path;
+
+        let path = common_storage_handler(ctx.logger, &storage)?;
+        new_device(path)
+    }
+}
+
+#[derive(Debug)]
+pub struct PmemHandler {}
+
+#[async_trait::async_trait]
+impl StorageHandler for PmemHandler {
+    #[instrument]
+    async fn create_device(
+        &self,
+        storage: Storage,
+        ctx: &mut StorageContext,
+    ) -> Result<StorageDeviceObject> {
+        // Retrieve the device for pmem storage
+        wait_for_pmem_device(ctx.sandbox, &storage.source).await?;
+
+        let path = common_storage_handler(ctx.logger, &storage)?;
+        new_device(path)
+    }
+}
